requirements.txt
# SurgicalDistro Backend - Python dependencies
# Install: pip install -r requirements.txt

# Core
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1

# Auth (JWT)
djangorestframework-simplejwt==5.5.1

# Database (PostgreSQL; optional for production)
psycopg2-binary==2.9.11

# PDF generation (invoices)
weasyprint==68.1

# Production server (e.g. PythonAnywhere, Linux)
gunicorn==25.1.0
